Set entities IDs to be of type Ling instead

+++ src/main/resources/db/migration/V1__create_topic_table.sql
@@ -5,7 +5,7 @@ CREATE SEQUENCE hibernate_sequence;
 DROP TABLE IF EXISTS topic;
 CREATE TABLE topic
 (
-    id                      INT  GENERATED ALWAYS AS IDENTITY,
+    id                      BIGINT  GENERATED ALWAYS AS IDENTITY,
     name                    VARCHAR  NOT NULL,
     description             VARCHAR  NOT NULL,
     PRIMARY KEY (id)
@@ -15,10 +15,10 @@ CREATE TABLE topic
 DROP TABLE IF EXISTS course;
 CREATE TABLE course
 (
-    id                      INT  GENERATED ALWAYS AS IDENTITY,
+    id                      BIGINT  GENERATED ALWAYS AS IDENTITY,
     name                    VARCHAR   NOT NULL,
     description             VARCHAR   NOT NULL,
-    topic_id                INT NOT NULL,
+    topic_id                BIGINT NOT NULL,
     PRIMARY KEY (id),
     CONSTRAINT fk_topic FOREIGN KEY (topic_id) REFERENCES topic(id)
 );
