+++ src/main/resources/db/migration/V2__create_lesson_table.sql
@@ -2,10 +2,10 @@
 DROP TABLE IF EXISTS lesson;
 CREATE TABLE lesson
 (
-    id                      INT  GENERATED ALWAYS AS IDENTITY,
+    id                      BIGINT  GENERATED ALWAYS AS IDENTITY,
     name                    VARCHAR   NOT NULL,
     description             VARCHAR   NOT NULL,
-    course_id               INT NOT NULL,
+    course_id               BIGINT NOT NULL,
     PRIMARY KEY(id),
     CONSTRAINT fk_course FOREIGN KEY (course_id) REFERENCES course(id)
 );
